+++ src/pages/sos/subtype.js
@@ -24,7 +24,7 @@ function RenderDate(props) {
     const { hasFocus, value } = props;
     const buttonElement = React.useRef(null);
     const rippleRef = React.useRef(null);
-   
+
 
     React.useLayoutEffect(() => {
         if (hasFocus) {
@@ -60,49 +60,57 @@ RenderDate.propTypes = {
 };
 
 const SubType = () => {
-
     const [rows, setRows] = React.useState(initialRows);
     const [loading, setLoading] = useState(false);
+    const Swal = require('sweetalert2')
+
 
     const Delete = React.useCallback(
         (id) => () => {
-            try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
-                const headers = { 'Authorization': AuthStr };
-                const apiUrl = `http://localhost:81/SosApp/emergency/admin/subType/${id}`;
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
-                // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
-                axios.delete(apiUrl, { headers })
-                    .then(response => {
-                        // If request is good...
-                        console.log(response.data.data);
-                        window.location.reload();
 
-                    })
-                    .catch((error) => {
-                        console.log('error ' + error);
-                    });
-                // return data;
-            } catch (error) {
-                // throw new Error(error);
-                console.error(error);
-                return error.response;
-            }
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    try {
+                        const token = localStorage.getItem('token')
+                        const AuthStr = 'Bearer '.concat(token);
+                        const headers = { 'Authorization': AuthStr };
+                        const apiUrl = `http://localhost:81/SosApp/emergency/admin/subType/${id}`;
+                        axios.delete(apiUrl, { headers })
+                            .then(response => {
+                                Swal.fire(
+                                    {
+                                        title: 'Deleted',
+                                        text: "Your file has been deleted.",
+                                        icon: 'warning',
+                                        confirmButtonText: 'Ok'
+                                    }
+                                ).then((result) => {
+                                    if (result.isConfirmed) {
+                                        window.location.reload();
+                                    }
+                                })
+                            })
+                            .catch((error) => {
+                                console.log('error ' + error);
+                            });
+                    } catch (error) {
+                        console.error(error);
+                        return error.response;
+                    }
+
+                }
+            })
         },
         [],
     );
-
-    const toggleAdmin = React.useCallback(
-        (id) => () => {
-            setRows((prevRows) =>
-                prevRows.map((row) =>
-                    row.id === id ? { ...row, isAdmin: !row.isAdmin } : row,
-                ),
-            );
-        },
-        [],
-    );
-
 
 
     const columns = [
@@ -111,7 +119,7 @@ const SubType = () => {
         {
             field: 'imageSubType', headerName: 'Image sub type', width: 300,
             renderCell: (params) => {
-                console.log(params);
+               
                 return (
                     <>
                         <img src={`data:image/jpeg;base64,${params.row.imageSubType}`} />
@@ -130,7 +138,7 @@ const SubType = () => {
                     icon={<DeleteIcon />}
                     label="Delete"
                     onClick={Delete(params.id)}
-                    showInMenu
+                    
                 />,
             ],
         },
@@ -141,16 +149,17 @@ const SubType = () => {
             setLoading(true);
 
             try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
+                const token = localStorage.getItem('token')
+                const AuthStr = 'Bearer '.concat(token);
                 const apiUrl = 'http://localhost:81/SosApp/emergency/subType';
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
+              
                 // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
                 await axios.get(apiUrl, { headers: { Authorization: AuthStr } })
                     .then(response => {
                         // If request is good...
-                        console.log(response.data.data);
+                      
                         setRows(response.data.data);
-                        console.log(rows);
+                      
                         setLoading(false);
 
                     })
@@ -161,8 +170,12 @@ const SubType = () => {
                     });
                 // return data;
             } catch (error) {
-                // throw new Error(error);
-                console.error(error);
+                Swal.fire({
+                    title: 'Error!',
+                    text: `${error.response.data.message}`,
+                    icon: 'error',
+                    confirmButtonText: 'Close'
+                  })
                 return error.response;
             } finally {
                 setLoading(false);
