+++ src/pages/sos/type.js
@@ -58,46 +58,53 @@ RenderDate.propTypes = {
 };
 
 const Type = () => {
-
     const [rows, setRows] = React.useState(initialRows);
     const [loading, setLoading] = useState(false);
-
+    const Swal = require('sweetalert2')
 
     const Delete = React.useCallback(
         (id) => () => {
-            try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
-                const headers = { 'Authorization': AuthStr };
-                const apiUrl = `http://localhost:81/SosApp/emergency/admin/type/${id}`;
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
-                // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
-                axios.delete(apiUrl, { headers })
-                    .then(response => {
-                        // If request is good...
-                        console.log(response.data.data);
-                        window.location.reload();
 
-                    })
-                    .catch((error) => {
-                        console.log('error ' + error);
-                    });
-                // return data;
-            } catch (error) {
-                // throw new Error(error);
-                console.error(error);
-                return error.response;
-            }
-        },
-        [],
-    );
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    try {
+                        const token = localStorage.getItem('token')
+                        const AuthStr = 'Bearer '.concat(token);
+                        const headers = { 'Authorization': AuthStr };
+                        const apiUrl = `http://localhost:81/SosApp/emergency/admin/type/${id}`;
+                        axios.delete(apiUrl, { headers })
+                            .then(response => {
+                                Swal.fire(
+                                    {
+                                        title: 'Deleted',
+                                        text: "Your file has been deleted.",
+                                        icon: 'warning',
+                                        confirmButtonText: 'Ok'
+                                    }
+                                ).then((result) => {
+                                    if (result.isConfirmed) {
+                                        window.location.reload();
+                                    }
+                                })
+                            })
+                            .catch((error) => {
+                                console.log('error ' + error);
+                            });
+                    } catch (error) {
+                        console.error(error);
+                        return error.response;
+                    }
 
-    const toggleAdmin = React.useCallback(
-        (id) => () => {
-            setRows((prevRows) =>
-                prevRows.map((row) =>
-                    row.id === id ? { ...row, isAdmin: !row.isAdmin } : row,
-                ),
-            );
+                }
+            })
         },
         [],
     );
@@ -110,7 +117,7 @@ const Type = () => {
         {
             field: 'imageType', headerName: 'Image type', width: 300,
             renderCell: (params) => {
-                console.log(params);
+              
                 return (
                     <>
                         <img src={`data:image/jpeg;base64,${params.row.imageType}`} />
@@ -129,7 +136,7 @@ const Type = () => {
                     icon={<DeleteIcon />}
                     label="Delete"
                     onClick={Delete(params.id)}
-                    showInMenu
+                    
                 />,
             ],
         },
@@ -140,18 +147,18 @@ const Type = () => {
             setLoading(true);
 
             try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
+                const token = localStorage.getItem('token')
+                const AuthStr = 'Bearer '.concat(token);
                 const apiUrl = 'http://localhost:81/SosApp/emergency/type';
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
-                // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
+
                 await axios.get(apiUrl, { headers: { Authorization: AuthStr } })
                     .then(response => {
                         // If request is good...
-                        console.log(response.data.data);
+                      
                         setRows(response.data.data);
                         setLoading(false);
 
-                        console.log(rows);
+                       
                     })
                     .catch((error) => {
                         console.log('error ' + error);
@@ -160,8 +167,12 @@ const Type = () => {
                     });
                 // return data;
             } catch (error) {
-                // throw new Error(error);
-                console.error(error);
+                Swal.fire({
+                    title: 'Error!',
+                    text: `${error.response.data.message}`,
+                    icon: 'error',
+                    confirmButtonText: 'Close'
+                  })
                 return error.response;
             } finally {
                 setLoading(false);
@@ -175,44 +186,44 @@ const Type = () => {
 
     return (
         <>
-          {loading ?
-            <>
-              <Loading />
-              <div className='opacity-20'>
-                <Grid container spacing={2}>
-                  <Grid item xs={6}>
-                    <p className='text-xl'><span><SosIcon className='mr-5' /></span>UserManagement</p>
-                  </Grid>
-                  <Grid item xs={6}>
-                    <DialogsSos />
-                  </Grid>
-                </Grid>
-                <div style={{ height: 400, width: '100%' }}>
-                  <DataGrid rows={rows} columns={columns} />
-                </div>
-              </div>
-            </>
-            :
-            <>
-              <Grid container spacing={2}>
-                <Grid item xs={6}>
-                  <p className='text-xl'><span><SosIcon className='mr-5' /></span>UserManagement</p>
-                </Grid>
-                <Grid item xs={6}>
-                  <DialogsSos />
-                </Grid>
-              </Grid>
-    
-              <div style={{ height: 400, width: '100%' }}>
-                <DataGrid rows={rows} columns={columns} />
-              </div>
-            </>
-    
-          }
-    
+            {loading ?
+                <>
+                    <Loading />
+                    <div className='opacity-20'>
+                        <Grid container spacing={2}>
+                            <Grid item xs={6}>
+                                <p className='text-xl'><span><SosIcon className='mr-5' /></span>UserManagement</p>
+                            </Grid>
+                            <Grid item xs={6}>
+                                <DialogsSos />
+                            </Grid>
+                        </Grid>
+                        <div style={{ height: 400, width: '100%' }}>
+                            <DataGrid rows={rows} columns={columns} />
+                        </div>
+                    </div>
+                </>
+                :
+                <>
+                    <Grid container spacing={2}>
+                        <Grid item xs={6}>
+                            <p className='text-xl'><span><SosIcon className='mr-5' /></span>UserManagement</p>
+                        </Grid>
+                        <Grid item xs={6}>
+                            <DialogsSos />
+                        </Grid>
+                    </Grid>
+
+                    <div style={{ height: 400, width: '100%' }}>
+                        <DataGrid rows={rows} columns={columns} />
+                    </div>
+                </>
+
+            }
+
         </>
-      );
-    }
+    );
+}
