+++ src/pages/users/roleManagement.js
@@ -15,6 +15,7 @@ import axios from 'axios';
 import Grid from '@mui/material/Grid';
 import AdminPanelSettingsIcon from '@mui/icons-material/AdminPanelSettings';
 import Loading from '../../shared/Loading/index';
+import Swal from 'sweetalert2'
 
 
 
@@ -57,59 +58,58 @@ RenderDate.propTypes = {
 };
 
 const RoleManagement = () => {
-
     const [rows, setRows] = React.useState(initialRows);
     const [loading, setLoading] = useState(false);
+    const Swal = require('sweetalert2')
 
 
     const Delete = React.useCallback(
         (id) => () => {
-            try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
-                const headers = { 'Authorization': AuthStr };
-                const apiUrl = `http://localhost:80/SosApp/accounts/admin/role/${id}`;
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
-                // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
-                axios.delete(apiUrl, { headers })
-                    .then(response => {
-                        // If request is good...
-                        console.log(response.data.data);
-                        window.location.reload();
 
-                    })
-                    .catch((error) => {
-                        console.log('error ' + error);
-                    });
-                // return data;
-            } catch (error) {
-                // throw new Error(error);
-                console.error(error);
-                return error.response;
-            }
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    try {
+                        const token = localStorage.getItem('token')
+                        const AuthStr = 'Bearer '.concat(token);
+                        const headers = { 'Authorization': AuthStr };
+                        const apiUrl = `http://localhost:80/SosApp/accounts/admin/role/${id}`;
+                        axios.delete(apiUrl, { headers })
+                            .then(response => {
+                                Swal.fire(
+                                    {
+                                        title: 'Deleted',
+                                        text: "Your file has been deleted.",
+                                        icon: 'warning',
+                                        confirmButtonText: 'Ok'
+                                    }
+                                ).then((result) => {
+                                    if (result.isConfirmed) {
+                                        window.location.reload();
+                                    }
+                                })
+                            })
+                            .catch((error) => {
+                                console.log('error ' + error);
+                            });
+                    } catch (error) {
+                        console.error(error);
+                        return error.response;
+                    }
+
+                }
+            })
         },
         [],
     );
 
-    const toggleAdmin = React.useCallback(
-        (id) => () => {
-            setRows((prevRows) =>
-                prevRows.map((row) =>
-                    row.id === id ? { ...row, isAdmin: !row.isAdmin } : row,
-                ),
-            );
-        },
-        [],
-    );
-
-    const duplicateUser = React.useCallback(
-        (id) => () => {
-            setRows((prevRows) => {
-                const rowToDuplicate = prevRows.find((row) => row.id === id);
-                return [...prevRows, { ...rowToDuplicate, id: Date.now() }];
-            });
-        },
-        [],
-    );
 
     const columns = [
         { field: 'name', headerName: 'Role', width: 1000 },
@@ -133,26 +133,30 @@ const RoleManagement = () => {
         const fetchData = async () => {
             setLoading(true);
             try {
-                const AuthStr = 'Bearer '.concat('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg');
+                const token = localStorage.getItem('token')
+                const AuthStr = 'Bearer '.concat(token);
                 const apiUrl = 'http://localhost:80/SosApp/accounts/admin/role';
-                const config = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIiwiZXhwIjoxNjg2Mzg5MDc0fQ.eEkGAKDc2HbUDWrngr4y19LYkyOnfLc10Ihhbh_KTzg';
-                // const { data } = await axios.get(apiUrl, { 'headers': { 'Authorization': AuthStr } });
                 await axios.get(apiUrl, { headers: { Authorization: AuthStr } })
                     .then(response => {
-                        // If request is good...
-                        console.log(response.data.data.getRoleList);
+              
+                 
                         setRows(response.data.data.getRoleList);
-                        console.log(rows);
+                       
                         setLoading(false);
                     })
                     .catch((error) => {
-                        console.log('error ' + error);
+                    
                         setLoading(false);
                     });
-                // return data;
+             
             } catch (error) {
-                // throw new Error(error);
-                console.error(error);
+             
+                Swal.fire({
+                    title: 'Error!',
+                    text: `${error.response.data.message}`,
+                    icon: 'error',
+                    confirmButtonText: 'Close'
+                  })
                 return error.response;
             } finally {
                 setLoading(false);
